# ARROW-5527: [C++] Uses Buffer/Builder in HashTable and MemoTable