… cpp/src/arrow/array-dict-test.cc
… cpp/src/arrow/util/hashing-test.cc
… cpp/src/parquet/encoding.cc