In order to prepare for the GroupBy operator, we need to cap memory usage of hash tables. This is done by passing a custom allocator to all HashTable/MemoTable and forces them to use Buffers for heap storage instead of std::vector and std::string (in the case of BinaryMemoTable).

There's a noted 5-10% performance penalty in the string benchmarks (negligible change for integers). The BuildStringDictionary benchmark pays a 30% penalty, due to Upsize and the memset. This overhead is mostly in memset, which can be controlled by a better default table size. One can validate by changing the default MemoTable values in the Hash kernel.

This patch hides the failure via DCHECK_OK internaly, the next step will be to bubble up the failing status to kernels, but this will require more careful tuning.

There's still improvement to be done, e.g. the CopyValues/CopyOffsets/GetDictionary can now be mostly zero-copy, as long as the MemoTable is also finalized/moved. I didn't change this contract because I think some consumers uses this class in an iterative fashion (the pandas stuff?). The final end goal is to take ownership of indices and values in a zero copy fashion for the future group-by on multiple keys.