@fsaintjacques 2019-07-12
@ursabot benchmark --suite-filter=arrow-compute-benchmark HEAD~4