@fsaintjacques 2019-07-29
@pitrou the DCHECK are there temporarily. I realize I was not clear in the PR description. My goal is a follow-up PR where the Status are bubbled up. This will touch all the hash kernels and some other parts. I wanted to minimized the diff to review.