@pitrou 2019-07-30
@fsaintjacques Is this still WIP?
Also, sorry, but you'll need to rebase following the LargeBinary merge...