@fsaintjacques 2019-07-30
@ursabot benchmark --suite-filter=arrow-compute-benchmark HEAD~4