@fsaintjacques 2019-07-30
@pitrou can you reproduce this benchmark regression on your local machine? I get around 0~2% here.