@pitrou 2019-07-31
Ok, I'm using clang 7.0 on Ubuntu 18.04.
```
$ taskset -c 3 ./build-release/release/arrow-compute-benchmark 
```

* Before:
```
--------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations
--------------------------------------------------------------------------
BuildDictionary                       2826753 ns    2825502 ns        260   1.38115GB/s
BuildStringDictionary                 5440800 ns    5436982 ns        126    55.528MB/s
UniqueInt64NoNulls/4194304/1024      11870135 ns   11862724 ns         59    2.6343GB/s
UniqueInt64NoNulls/4194304/10240     15810591 ns   15801368 ns         44   1.97768GB/s
UniqueInt64WithNulls/4194304/1024    14531716 ns   14522688 ns         48   2.15181GB/s
UniqueInt64WithNulls/4194304/10240   18381640 ns   18371398 ns         38   1.70101GB/s
UniqueString10bytes/4194304/1024     63875366 ns   63852270 ns         11   626.446MB/s
UniqueString10bytes/4194304/10240    96171205 ns   96144417 ns          7   416.041MB/s
UniqueString100bytes/4194304/1024   202963923 ns  202852292 ns          3   1.92566GB/s
UniqueString100bytes/4194304/10240  269573917 ns  269416498 ns          3   1.44989GB/s
UniqueUInt8NoNulls/4194304/200          24104 us      24098 us         29   165.986MB/s
UniqueUInt8WithNulls/4194304/200        26359 us      26353 us         26   151.785MB/s
```

* After:
```
--------------------------------------------------------------------------
Benchmark                                   Time           CPU Iterations
--------------------------------------------------------------------------
BuildDictionary                       2848175 ns    2846933 ns        239   1.37075GB/s
BuildStringDictionary                 5632237 ns    5627822 ns        124    53.645MB/s
UniqueInt64NoNulls/4194304/1024      34443000 ns   34431143 ns         20   929.391MB/s
UniqueInt64NoNulls/4194304/10240     14430863 ns   14421642 ns         48   2.16688GB/s
UniqueInt64WithNulls/4194304/1024    40027697 ns   40014341 ns         17   799.713MB/s
UniqueInt64WithNulls/4194304/10240   18194800 ns   18184094 ns         39   1.71853GB/s
UniqueString10bytes/4194304/1024     71903582 ns   71881712 ns         10    556.47MB/s
UniqueString10bytes/4194304/10240   101814901 ns  101782233 ns          7   392.996MB/s
UniqueString100bytes/4194304/1024   208658610 ns  208539334 ns          3   1.87315GB/s
UniqueString100bytes/4194304/10240  273170198 ns  273014066 ns          3   1.43079GB/s
UniqueUInt8NoNulls/4194304/200          24098 us      24092 us         29   166.031MB/s
UniqueUInt8WithNulls/4194304/200        26168 us      26162 us         27   152.894MB/s
```