@pitrou 2019-07-31
Also:
```
$ taskset -c 3 ./build-release/release/arrow-builder-benchmark --benchmark_filter=Dict
```

* Before:
```
---------------------------------------------------------------------------
Benchmark                                    Time           CPU Iterations
---------------------------------------------------------------------------
BuildInt64DictionaryArrayRandom      558981604 ns  558809559 ns          1   458.117MB/s
BuildInt64DictionaryArraySequential  514187491 ns  514045372 ns          1   498.011MB/s
BuildInt64DictionaryArraySimilar     745244503 ns  744974263 ns          1   343.636MB/s
BuildStringDictionaryArray          1311752166 ns 1311277919 ns          1    260.46MB/s
```
* After:
```
---------------------------------------------------------------------------
Benchmark                                    Time           CPU Iterations
---------------------------------------------------------------------------
BuildInt64DictionaryArrayRandom      663034499 ns  662828769 ns          1   386.223MB/s
BuildInt64DictionaryArraySequential  821072696 ns  820888109 ns          1   311.857MB/s
BuildInt64DictionaryArraySimilar    1122597093 ns 1122034504 ns          1   228.157MB/s
BuildStringDictionaryArray          1607663812 ns 1606582405 ns          1   212.585MB/s
```