@pitrou 2019-07-31
So there's clearly a regression here.