@pitrou 2019-07-31
Hmm, some of the benchmarks seem very sensitive to the initial hash table capacity. On my CPU, the numbers 512 and 2048 seem cursed, but 1024 is better. Perhaps some cache aliasing issue, or something.