@pitrou 2019-07-31
Perhaps we should step back a bit. It's difficult to write benchmarks for hash tables, especially micro-benchmarks, since you can easily stress corner cases and get a false sense of performance improvements or regressions.

So the benchmark number variations here are not necessarily an issue in the hash table code.