@pitrou 2019-07-31
>  I'm not entirely satisfied with the end-result.

With the code or with the performance? As I said I think we can live with the performance changes for now, and revisit later.