@wesm 2019-07-31
The perf regression definitely isn't good. I can't look at this today but I might like to poke at this patch a little bit. Is there any time urgency to merge this?