@wesm 2019-08-01
This ArraySimilar benchmark is worrisome

after:

```
---------------------------------------------------------------------------------------
Benchmark                                                Time           CPU Iterations
---------------------------------------------------------------------------------------
BuildInt64DictionaryArraySimilar/min_time:5.000  881550805 ns  881549006 ns          8   290.398MB/s
```
Flamegraph: https://svgshare.com/i/EP2.svg

before:

```
---------------------------------------------------------------------------------------
Benchmark                                                Time           CPU Iterations
---------------------------------------------------------------------------------------
BuildInt64DictionaryArraySimilar/min_time:5.000  630083157 ns  630082451 ns         11   406.296MB/s
```

Flamegraph: https://svgshare.com/i/EN8.svg

I'm digging a bit to see if I can learn anything more